+++ MovieEncoder/BackupDiskHandBrakeJob.cs
@@ -10,13 +10,15 @@ namespace MovieEncoder
     {
         private readonly HandBrakeService handBrakeService;
         private readonly string driveName;
+        private readonly bool backupAll;
 
         public override string JobName => "Scan Disk " + driveName;
 
-        public BackupDiskHandBrakeJob(HandBrakeService handBrakeService, string driveName)
+        public BackupDiskHandBrakeJob(HandBrakeService handBrakeService, string driveName, bool backupAll)
         {
             this.handBrakeService = handBrakeService;
             this.driveName = driveName;
+            this.backupAll = backupAll;
         }
 
         public override bool RunJob(JobQueue jobRunner)
@@ -29,7 +31,7 @@ namespace MovieEncoder
                 diskTitles.Sort((o1, o2) => {
                     if (o1.Seconds > o2.Seconds)
                     {
-                        return 1;
+                        return -1;
                     }
                     else if (o1.Seconds == o2.Seconds)
                     {
@@ -37,19 +39,32 @@ namespace MovieEncoder
                     }
                     else
                     {
-                        return -1;
+                        return 1;
                     }
                 });
                 DiskTitle mainTitle = diskTitles.Find((o) => o.MainMovie == true);
                 if (mainTitle != null)
                 {
                     diskTitles.Remove(mainTitle);
-                    diskTitles.Insert(0, mainTitle);
+                    if (!backupAll)
+                    {
+                        diskTitles.Clear();
+                        diskTitles.Add(mainTitle);
+                    }
+                    else
+                    {
+                        diskTitles.Insert(0, mainTitle);
+                    }
+                } else if (!backupAll) {
+                    // pick the 1st 
+                    mainTitle = diskTitles[0];
+                    diskTitles.Clear();
+                    diskTitles.Add(mainTitle);
                 }
 
                 foreach (DiskTitle title in diskTitles)
                 {
-                    string cleanTitle = GetMovieTitle(title) + String.Format("_t{0:00}.mkv", n++);
+                    string cleanTitle = GetMovieTitle(title) + String.Format("_t{0:00}.{1}", n++, handBrakeService.MovieOutputType == HandBrakeService.OutputType.MP4 ? "mp4" : "mkv");
                     jobRunner.AddJob(new EncodeMovieJob(handBrakeService, driveName, cleanTitle, title.TitleIndex, false));
                 }
                 return true;