+++ MovieEncoder/BackupDiskMakeMKVJob.cs
@@ -116,42 +116,5 @@ namespace MovieEncoder
                 }
             }
         }
-
-        private static int ComparDiskTitleByHorizontalResolution(DiskTitle x, DiskTitle y)
-        {
-            if (x == null)
-            {
-                if (y == null)
-                {
-                    return 0;
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-            else
-            {
-                if (y == null)
-                {
-                    return 1;
-                }
-                else
-                {
-                    if (x.HorizontalResolution < y.HorizontalResolution)
-                    {
-                        return -1;
-                    }
-                    else if (x.HorizontalResolution == y.HorizontalResolution)
-                    {
-                        return 0;
-                    }
-                    else
-                    {
-                        return 1;
-                    }
-                }
-            }
-        }
     }
 }