+++ MovieEncoder/EncodeMovieJob.cs
@@ -30,7 +30,7 @@ namespace MovieEncoder
         private readonly HandBrakeService handBrakeService;
         private readonly bool keepFiles;
         private readonly int titleIndex;
-        private string movieTitle;
+        private readonly string movieTitle;
 
         public EncodeMovieJob(HandBrakeService handBrakeService, string inputFilePath, int titleIndex, bool keepFiles)
         {
@@ -80,67 +80,80 @@ namespace MovieEncoder
             }
 
             progressReporter.CurrentTask = "Getting titles from file using Handbrake";
-            //List<DiskTitle> diskTitles = handBrakeService.Scan(InputFileName, titleIndex, false, progressReporter);
-            //if (diskTitles.Count > 0)
+            // if input is a directory, then assume its a drive
+            string outputFile;
+            if (Directory.Exists(InputFileName) && movieTitle != null)
             {
-                // only take 1st
-                //DiskTitle diskTitle = diskTitles[0];
+                // generate the output file based on the scan output
+                outputFile = Path.Combine(OutputDir, movieTitle);
+            } else
+            {
+                // skip if the output file already exists
+                outputFile = Path.Combine(OutputDir, Path.GetFileName(InputFileName));
+            }
 
-                // if input is a directory, then assume its a drive
-                string outputFile;
-                if (Directory.Exists(InputFileName) && movieTitle != null)
+            outputFile = RenameExtension(outputFile);
+
+            if (File.Exists(outputFile)) {
+                throw new JobException($"Encoding may have already been run. Output file '{outputFile}' already exists.");
+            }
+
+            progressReporter.CurrentTask = "Start encoding using Handbrake";
+            if (handBrakeService.Encode(InputFileName, titleIndex, outputFile, progressReporter))
+            {
+                // if the output file exists, remove the input file
+                if (File.Exists(outputFile))
                 {
-                    // generate the output file based on the scan output
-                    outputFile = Path.Combine(OutputDir, movieTitle);
+                    // always delete from the source directory.
+                    string sourceDirectory = handBrakeService.HandBrakeSourceDir;
+                    if (sourceDirectory.EndsWith(Path.DirectorySeparatorChar.ToString()))
+                    {
+                        sourceDirectory = sourceDirectory.Substring(0, sourceDirectory.Length - 1);
+                    }
+                    if (movieTitle == null && (!keepFiles || Path.GetDirectoryName(InputFileName).StartsWith(sourceDirectory)))
+                    {
+                        // Delete the source file
+                        File.Delete(InputFileName);
+                        // check for source directory
+                        string directory = Path.GetDirectoryName(InputFileName);
+                        while (!directory.Equals(sourceDirectory))
+                        {
+                            // if the name of the directory matches the file name and
+                            // the directory is empty, delete it.
+                            if (InputFileName.StartsWith(directory))
+                            {
+                                if (Directory.GetFiles(directory).Length == 0)
+                                {
+                                    Directory.Delete(directory);
+                                }
+                            }
+                            directory = Path.GetDirectoryName(directory);
+                        }
+                    }
+                    return true;
                 } else
                 {
-                    // skip if the output file already exists
-                    outputFile = Path.Combine(OutputDir, Path.GetFileName(InputFileName));
-                }
-                if (File.Exists(outputFile)) {
-                    throw new JobException($"Encoding may have already been run. Output file '{outputFile}' already exists.");
-                }
-
-                progressReporter.CurrentTask = "Start encoding using Handbrake";
-                if (handBrakeService.Encode(InputFileName, titleIndex, outputFile, progressReporter))
-                {
-                    // if the output file exists, remove the input file
-                    if (File.Exists(outputFile))
-                    {
-                        // always delete from the source directory.
-                        string sourceDirectory = handBrakeService.HandBrakeSourceDir;
-                        if (sourceDirectory.EndsWith(Path.DirectorySeparatorChar.ToString()))
-                        {
-                            sourceDirectory = sourceDirectory.Substring(0, sourceDirectory.Length - 1);
-                        }
-                        if (movieTitle == null && (!keepFiles || Path.GetDirectoryName(InputFileName).StartsWith(sourceDirectory)))
-                        {
-                            // Delete the source file
-                            File.Delete(InputFileName);
-                            // check for source directory
-                            string directory = Path.GetDirectoryName(InputFileName);
-                            while (!directory.Equals(sourceDirectory))
-                            {
-                                // if the name of the directory matches the file name and
-                                // the directory is empty, delete it.
-                                if (InputFileName.StartsWith(directory))
-                                {
-                                    if (Directory.GetFiles(directory).Length == 0)
-                                    {
-                                        Directory.Delete(directory);
-                                    }
-                                }
-                                directory = Path.GetDirectoryName(directory);
-                            }
-                        }
-                        return true;
-                    } else
-                    {
-                        throw new Exception($"Missing output file {outputFile}. Not removing original file.");
-                    }
+                    throw new Exception($"Missing output file {outputFile}. Not removing original file.");
                 }
             }
             return false;
+        }
+
+        private string RenameExtension(string outputFile)
+        {
+            string ext = ".mkv";
+            if (handBrakeService.MovieOutputType == HandBrakeService.OutputType.MP4)
+            {
+                ext = ".mp4";
+            }
+            
+            if (!outputFile.EndsWith(ext))
+            {
+                // rename
+                int idx = outputFile.LastIndexOf('.');
+                outputFile = outputFile.Substring(0, idx) + ext;
+            }
+            return outputFile;
         }
     }
 }