+++ MovieEncoder/Job.cs
@@ -15,6 +15,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
+using System.Diagnostics;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text;
@@ -23,6 +24,7 @@ using System.Windows.Data;
 
 namespace MovieEncoder
 {
+    [DebuggerDisplay("JobName = {JobName}")]
     public abstract class Job : INotifyPropertyChanged
     {
         private double _maxProgress = 100;