+++ CHANGELOG.md
@@ -4,9 +4,10 @@ Backup and re-encode bulk movies using projects MakeMKV and HandBrake
 
 ## TODO
 - Check for available disk space before backing up movies.
-- Support MP4 output
-- Support to keep HandBrake source files
+- Support to keep HandBrake source files.
   - Possible cache file / db to note which files have been enoded so not to attempt on startup.
+- Save process output to file.
+- Add support to filter out movies in handbrake backups.
 
 ## [1.0.1] - Unreleased
 ### Added
@@ -17,12 +18,16 @@ Backup and re-encode bulk movies using projects MakeMKV and HandBrake
 - Log information from MakeMKV scan.
 - Added lines between job log entries.
 - Add support for HandBrake to backup disks.
+- Support MP4 output.
+- Added flag for forced subtitles.
 
 ### Changed
 - Changed how multiple backup job ran. Now split into a scan job and a backup job.
 - Leave the performance page available after stopping the service.
 - Code cleanups in progress code.
 - Change color of error jobs in job list.
+- Enhanced Handbrake progress.
+- Clear log without clearing the current task log entries.
 
 ## [1.0.0] - 2020-07-04
 ### Added